# internal/web/assets/openapi.yaml
openapi: 3.0.3
info:
  title: PasarGuard Panel API
  version: "1.0.0"
  description: >
    API پنل اصلی برای مدیریت چند نود، مشتری، پلن، اشتراک و سهمیه.
    این API سطح یکپارچه‌سازی برای ربات فروش است. پنل منطق پول/کیف‌پول ندارد؛
    فقط بایت/زمان/وضعیت را مدیریت و مصرف/overage را expose می‌کند.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /sub/{token}:
    get:
      tags: [system]
      summary: Public subscription page (HTML) for the customer — config + quota/expiry
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: HTML page }
        "404": { description: Not found }
...
  /api/v1/customers/{id}/subscriptions:
    get:
      tags: [subscriptions]
      summary: List a customer's subscriptions
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Subscription" }
    post:
      tags: [subscriptions]
      summary: Provision a subscription (creates a tenant on a node)
      description: >
        یک tenant روی نود انتخابی می‌سازد و کلید API مشتری را **فقط یک‌بار** برمی‌گرداند.
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id, node_id]
              properties:
                plan_id: { type: string }
                node_id: { type: string }
                credit_limit_bytes: { type: integer, format: int64 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionResponse" }
        "502": { $ref: "#/components/responses/NodeError" }
...
  /api/v1/subscriptions/{id}:
    delete:
      tags: [subscriptions]
      summary: Delete a subscription (deprovisions the tenant from its node)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    NodeError:
      description: خطا در ارتباط با نود
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Subscription:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        plan_id: { type: string }
        node_id: { type: string }
        node_tenant_id: { type: string }
        status: { type: string, enum: [active, suspended, expired] }
        period_id: { type: integer }
        start_at: { type: integer, format: int64 }
        end_at: { type: integer, format: int64 }
        quota_bytes: { type: integer, format: int64 }
        used_bytes: { type: integer, format: int64 }
        credit_limit_bytes: { type: integer, format: int64 }
        notified_threshold: { type: integer }
        created_at: { type: integer, format: int64 }
    SubscriptionResponse:
      type: object
      properties:
        subscription: { $ref: "#/components/schemas/Subscription" }
        api_key: { type: string, description: "کلید مشتری — فقط همین یک‌بار نمایش داده می‌شود" }
        node_address: { type: string }
        sub_token: { type: string, description: "توکن صفحه‌ی عمومی اشتراک (/sub/{token})" }
...